requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
greenlet==1.1.2
idna==3.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.30
SQLAlchemy-Utils==0.38.2
urllib3==1.26.8
wincertstore==0.2

Pillow~=9.0.0
numpy~=1.21.5
imutils~=0.5.4
scipy~=1.7.3
gTTS~=2.2.3
